+++ src/main/java/org/money/web/UserContext.java
@@ -5,17 +5,17 @@ package org.money.web;
  * @since 2022/6/24 15:38
  */
 public class UserContext {
-    private static final ThreadLocal<Integer> userId = new ThreadLocal<>();
+    private static final ThreadLocal<Long> userId = new ThreadLocal<>();
 
     public static void remove() {
         userId.remove();
     }
 
-    public static void setUserId(Integer userId) {
+    public static void setUserId(Long userId) {
         UserContext.userId.set(userId);
     }
 
-    public static Integer getUserId() {
+    public static Long getUserId() {
         return userId.get();
     }
 }
